+++ packages/firestore/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Unreleased
+- [changed] Snapshot listeners (with the `includeMetadataChanges` option
+  enabled) now receive an event with `snapshot.metadata.fromCache` set to
+  `true` if the SDK loses its connection to the backend. A new event with
+  `snapshot.metadata.fromCache` set to false will be raised once the
+  connection is restored and the query is in sync with the backend again.
 - [changed] A "Could not reach Firestore backend." message will be
   logged when the initial connection to the Firestore backend fails.
 - [changed] A "Using maximum backoff delay to prevent overloading the